content fitting textareas, code cleanup
