+++ src/App.js
@@ -28,7 +28,7 @@ import ConfirmDetails from './pages/confirmDetails/confirmDetails';
 import BatchChange from './pages/batchChange/batchChange';
 import ProtectedRoute from './components/protectedRoute/protectedRoute';
 import PageNotFound from './pages/pageNotFound/pageNotFound';
-import ApplicationQuestions from './pages/scholarship/application_questions';
+import ApplicationQuestions from './pages/applicationQuestions/applicationQuestions';
 import TaxDetails from './pages/taxDetails/tax-details';
 import './App.css';
 
@@ -103,8 +103,10 @@ function App() {
 									<SelectionPage />
 								</ProtectedRoute>
 							}/>
-							<Route path='/scholarships/application_questions' element={
-								<ApplicationQuestions />
+							<Route path='/scholarships/application-questions' element={
+								<ProtectedRoute>
+									<ApplicationQuestions />
+								</ProtectedRoute>
 							}/>
 							<Route path = '/tax-details' element={<TaxDetails />}></Route>
 							<Route path='*' element={<PageNotFound />} />
